Programmatic analysis is working / working again
As forked from https://github.com/jensengroup/RegioSQM, script
regiosqm.py offered only the generation of input files to perform
the computations with MOPAC.  The instruction how to analyse the
results programmatically in the README.md did not match with the
design of the script and local work required the script's syntax
to be changed a little to enable its action.  Reading the git
log, the script used to prepare (and submit) the publication was
modified further; there is one release tag 1.0 by Sep 12, 2017
by Jan H. Jensen, but the doi was added by Oct 15, 2017, and the
internal version counter was incremented to 1.1 on Feb 10, 2018 --
while the publication 'as such' was accepted on Nov 10, 2017.

These problems are addressed by this commit as much as the script
may be used to be run locally for results to be compared with
results offered by the dedicated project site, regiosqm.org.  If
found to yield the results in agreement with the website, the
scripts should be revised; at present they may be used with legacy
Python 2.7.17 only.  Already the print instructions prevent their
deployment in current Python 3.x.