Added benchmark files about RegioSQM's results.
The Jensen group documented RegioSQM's work in the SI with 535
test structures; these SMILES are now provided as verbatim copy.
To probe if changes of the scripts affect the outcome, the mokka
test set (the first least conformationally flexible representative
about the first 36 EAS groups) is provided -- SMILES file,
conformational table, summary of the results; the later file then
may be used for a faster probe (by difference view) than running
the analysis over all test data and their conformations.

By visual comparison, results of this sub-set of data scrutinized
by the current set of scripts -- still working with Python 2.7.17
only -- visually match 1:1 the results published in the SI of the
seminal paper.