Sharp transition to Python 3 by 2to3-2.7
All five scripts were submitted to a translation into the syntax of
Python 3 by 2to3 (version 3.8.3-2 in Debian).  Script regiosqm.py
was altered further,
+ restoring the previous import statements in lieu of the suggested
  form in pattern of 'from . import protonate as prot'
+ altering the syntax of the clause checking for the occurence of
  an hydrogen transfer (currently lines #79 onward)

Under Python 3.8.4rc1 with openbabel (3.1.0), rdkit (201909.1-4.1),
the scripts successfully create conformers, write their .sdf and
MOPAC input files, and populate the table of conformation.  MOPAC
(20.173L) in turn is capable to perform the computations including
the report about the refined geometry and heat of formation.

TODO: Missing is the extraction of the heat of formation and output
as annotated .svg.