First set of scripts of regiosqm working with Python3
After multiple attempts this commit saves the first set of scripts
interacting smoothly -- from the list of SMILES in a file, till
and including the generation of the visual maps as .svg maps --
with Python3.  Tested with version Python (3.8.4rc1, Jul 1, 2020),
obabel (3.1.0, Jun 9, 2020) and RDKit (2019.09.1) as provided by
Linux Debian 10 buster / branch testing.