Streamline RegioSQM replicate input generation
Once pdftotext exctracted the text of the sections relevant to
replicate RegioSQM's work, it is now up to one script to prepare
the input files.

In case of missing reference file 'compounds_smiles.csv' with
compound numbers and SMILES encoded structures, the script will
both inform the user, and stop processing.