bug fix indices_display.py, display all atom indices
The previous form skipped the output of atom indices for the last
entry of the EAS group.  The cause, a wrong consecution of events
and an unneccessary nested for-loop, was quickly identified by
Ivan Tubert-Brohman, subscriber to the RDKit mail discussion list,
to whom equally the credit to the change of this commit is due.