in .svg, move ellipses behind the formula
The change of previous commits yielded .svg with formulae partially
behind the ellipses.  It seems better to place the ellipses behind
the strokes representing bonds, however, which is addressed here.

Equally, the inner documentation, especially in batch_regiosqm.py,
was revised to be easier to follow.  The parameter file written
now equally states the SMILES list input file accessed.